+++ crates/tzlint_rules/src/rules/mod.rs
@@ -6,6 +6,7 @@ pub mod max_ten;
 pub mod no_doubled_joshi;
 pub mod no_exclamation_question_mark;
 pub mod no_hankaku_kana;
+pub mod no_mix_dearu_desumasu;
 pub mod no_mixed_zenkaku_hankaku_alphabet;
 pub mod no_nfd;
 pub mod no_todo;